backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlmodel==0.0.16
faster-whisper==1.0.0
python-multipart==0.0.9
transformers==4.39.0
torch==2.2.1
ffmpeg-python==0.2.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
